associationMatrix Helper Functions

These objects contain a number of settings and functions for associationMatrix.

computeStatistic_t(var1, var2, conf.level = 0.95, var.equal = "test",
  ...)

computeStatistic_r(var1, var2, conf.level = 0.95, ...)

computeStatistic_f(var1, var2, conf.level = 0.95, ...)

computeStatistic_chisq(var1, var2, conf.level = 0.95, ...)

computeEffectSize_d(var1, var2, conf.level = 0.95, var.equal = "test",
  ...)

computeEffectSize_r(var1, var2, conf.level = 0.95, ...)

computeEffectSize_etasq(var1, var2, conf.level = 0.95, ...)

computeEffectSize_omegasq(var1, var2, conf.level = 0.95, ...)

computeEffectSize_v(var1, var2, conf.level = 0.95, bootstrap = FALSE,
  samples = 5000, ...)

Arguments

var1
One of the two variables for which to compute a statistic or effect size

var2
The other variable for which to compute the statistic or effect size

conf.level
The confidence for the confidence interval for the effect size

var.equal
Whether to test for equal variances ('test'), assume equality ('yes'), or assume unequality ('no'). See [userfriendlyscience::meanDiff()] for more information.

…
Any additonal arguments are sometimes used to specify exactly how statistics and effect sizes should be computed.

bootstrap
Whether to bootstrap to estimate the confidence interval for Cramer's V. If FALSE, the Fisher's Z conversion is used.

samples
If bootstrapping, the number of samples to generate (of course, more samples means more accuracy and longer processing time).

Value

associationMatrixStatDefaults and associationMatrixESDefaults contain the default functions from computeStatistic and computeEffectSize that are called (see the help file for associationMatrix for more details).

The other functions return an object with the relevant statistic or effect size, with a confidence interval for the effect size.

For computeStatistic, this object always contains:

statistic

The relevant statistic

statistic.type

The type of statistic

parameter

The degrees of freedom for this statistic

p.raw

The p-value of this statistic for NHST

And in addition, it often contains (among other things, sometimes):
object

The object from which the statistics are extracted

For computeEffectSize, this object always contains:
es

The point estimate for the effect size

esc.type

The type of effect size

ci

The confidence interval for the effect size

And in addition, it often contains (among other things, sometimes):
object

The object from which the effect size is extracted

See also

Examples

computeStatistic_f(Orange$Tree, Orange$circumference)

#> $object
#> Call:
#>    stats::aov(formula = dependent ~ factor)
#> 
#> Terms:
#>                    factor Residuals
#> Sum of Squares   11840.86 100525.43
#> Deg. of Freedom         4        30
#> 
#> Residual standard error: 57.88651
#> Estimated effects are balanced
#> 
#> $statistic
#> [1] 0.8834225
#> 
#> $statistic.type
#> [1] "f"
#> 
#> $parameter
#> [1]  4 30
#> 
#> $p.raw
#> [1] 0.4856567
#> 

computeEffectSize_etasq(Orange$Tree, Orange$circumference)

#> $realConfidence
#> [1] 0.9
#> 
#> $object.aov
#> Call:
#>    stats::aov(formula = dependent ~ factor)
#> 
#> Terms:
#>                    factor Residuals
#> Sum of Squares   11840.86 100525.43
#> Deg. of Freedom         4        30
#> 
#> Residual standard error: 57.88651
#> Estimated effects are balanced
#> 
#> $es
#> [1] 0.1053773
#> 
#> $es.type
#> [1] "etasq"
#> 
#> $object
#> $object$Lower.Limit.Proportion.of.Variance.Accounted.for
#> [1] 0
#> 
#> $object$Probability.Less.Lower.Limit
#> [1] 0
#> 
#> $object$Upper.Limit.Proportion.of.Variance.Accounted.for
#> [1] 0.1947565
#> 
#> $object$Probability.Greater.Upper.Limit
#> [1] 0.05
#> 
#> $object$Actual.Coverage
#> [1] 0.95
#> 
#> 
#> $ci
#> [1] 0.0000000 0.1947565
#>