Confidence intervals for Omega Squared

This function used the MBESS function [MBESS::conf.limits.ncf()] and [MBESS::convert.ncf.to.omegasq()] to compute the point estimate and confidence interval for Omega Squared (which have been lifted out of MBESS to avoid importing the whole package)

confIntOmegaSq(var1, var2, conf.level = 0.95)

# S3 method for confIntOmegaSq
print(x, ..., digits = 2)

Arguments

var1, var2
The two variables: one should be a factor (or will be made a factor), the other should have at least interval level of measurement. If none of the variables is a factor, the function will look for the variable with the least unique values and change it into a factor.

conf.level
Level of confidence for the confidence interval.

x, digits, …
Respectively the object to print, the number of digits to round to, and any additonal arguments to pass on to the `print` function.

Value

A confIntOmegaSq object is returned, with as elements:

input

The input arguments

intermediate

Objects generated while computing the output

output

The output of the function, consisting of:

output$es

The point estimate

output$ci

The confidence interval

Note

Formula 16 in Steiger (2004) is used for the conversion in convert.ncf.to.omegasq.

References

Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9(2), 164-82. https://doi.org/10.1037/1082-989X.9.2.164

Examples

confIntOmegaSq(mtcars$mpg, mtcars$cyl);

#> Omega squared: 95% CI = [.51; .81], point estimate = .71