The distribution of Omega Squared

These functions use some conversion to and from the F distribution to provide the Omega Squared distribution.

pomegaSq(q, df1, df2, populationOmegaSq = 0, lower.tail = TRUE)

qomegaSq(p, df1, df2, populationOmegaSq = 0, lower.tail = TRUE)

romegaSq(n, df1, df2, populationOmegaSq = 0)

domegaSq(x, df1, df2, populationOmegaSq = 0)

Arguments

df1, df2
Degrees of freedom for the numerator and the denominator, respectively.

populationOmegaSq
The value of Omega Squared in the population; this determines the center of the Omega Squared distribution. This has not been implemented yet in this version of userfriendlyscience. If anybody has the inverse of convert.ncf.to.omegasq for me, I'll happily integrate this.

lower.tail
logical; if TRUE (default), probabilities are the likelihood of finding an Omega Squared smaller than the specified value; otherwise, the likelihood of finding an Omega Squared larger than the specified value.

p
Vector of probabilites (p-values).

n
Desired number of Omega Squared values.

x, q
Vector of quantiles, or, in other words, the value(s) of Omega Squared.

Value

domegaSq gives the density, pomegaSq gives the distribution function, qomegaSq gives the quantile function, and romegaSq generates random deviates.

Details

The functions use convert.omegasq.to.f and convert.f.to.omegasq to provide the Omega Squared distribution.

See also

Examples

### Generate 10 random Omega Squared values
romegaSq(10, 66, 3);

#>  [1]  0.3771096  0.2769963  0.8639086 -1.2657636  0.5854294 -0.3364410
#>  [7] -1.1854040 -0.6081699  0.5009081  0.1671413

### Probability of findings an Omega Squared
### value smaller than .06 if it's 0 in the population
pomegaSq(.06, 66, 3);

#> [1] 0.4280867